Move function to date.js
