+++ src/core/components/courses/SlotContainer.vue
@@ -64,7 +64,8 @@ import SlotEditionModal from '@components/courses/SlotEditionModal';
 import SlotCreationModal from '@components/courses/SlotCreationModal';
 import { NotifyNegative, NotifyWarning, NotifyPositive } from '@components/popup/notify';
 import { E_LEARNING } from '@data/constants';
-import { formatQuantity, formatDate } from '@helpers/utils';
+import { formatQuantity } from '@helpers/utils';
+import { formatDate } from '@helpers/date';
 import { frAddress } from '@helpers/vuelidateCustomVal';
 import moment from '@helpers/moment';
 import { courseMixin } from '@mixins/courseMixin';
