+++ src/modules/client/pages/customers/Subscriptions.vue
@@ -142,7 +142,8 @@ import ResponsiveTable from '@components/table/ResponsiveTable';
 import { NotifyPositive, NotifyWarning, NotifyNegative } from '@components/popup/notify';
 import { REQUIRED_LABEL, DOC_EXTENSIONS } from '@data/constants';
 import { bic, iban } from '@helpers/vuelidateCustomVal';
-import { getLastVersion, formatDate } from '@helpers/utils';
+import { getLastVersion } from '@helpers/utils';
+import { formatDate } from '@helpers/date';
 import FundingGridTable from 'src/modules/client/components/table/FundingGridTable';
 import { customerMixin } from 'src/modules/client/mixins/customerMixin';
 import { subscriptionMixin } from 'src/modules/client/mixins/subscriptionMixin';
