+++ src/modules/client/pages/ni/auxiliaries/Directory.vue
@@ -46,8 +46,8 @@ import TableList from '@components/table/TableList';
 import DirectoryHeader from '@components/DirectoryHeader';
 import { NotifyPositive, NotifyNegative, NotifyWarning } from '@components/popup/notify';
 import { DEFAULT_AVATAR, AUXILIARY, AUXILIARY_ROLES, REQUIRED_LABEL, CIVILITY_OPTIONS, HR_SMS } from '@data/constants';
-import { formatIdentity, formatPhoneForPayload, removeDiacritics, sortStrings, formatDate } from '@helpers/utils';
-import { ascendingSort } from '@helpers/date';
+import { formatIdentity, formatPhoneForPayload, removeDiacritics, sortStrings } from '@helpers/utils';
+import { formatDate, ascendingSort } from '@helpers/date';
 import { userMixin } from '@mixins/userMixin';
 import { validationMixin } from '@mixins/validationMixin';
 import AuxiliaryCreationModal from 'src/modules/client/components/auxiliary/AuxiliaryCreationModal';
